+++ src/main/java/com/chuseok22/stocktrader/common/core/constants/Region.java
@@ -1,0 +1,5 @@
+package com.chuseok22.stocktrader.common.core.constants;
+
+public enum Region {
+  KR, US
+}
