+++ src/main/java/com/chuseok22/stocktrader/common/infrastructure/repository/RecommendationJpaRepository.java
@@ -1,6 +1,6 @@
-package com.chuseok22.stocktrader.infrastructure.repository;
+package com.chuseok22.stocktrader.common.infrastructure.repository;
 
-import com.chuseok22.stocktrader.infrastructure.entity.Recommendation;
+import com.chuseok22.stocktrader.common.infrastructure.entity.Recommendation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecommendationJpaRepository extends JpaRepository<Recommendation, Long> {
