+++ src/main/java/com/chuseok22/stocktrader/common/infrastructure/repository/StockJpaRepository.java
@@ -1,6 +1,6 @@
-package com.chuseok22.stocktrader.infrastructure.repository;
+package com.chuseok22.stocktrader.common.infrastructure.repository;
 
-import com.chuseok22.stocktrader.infrastructure.entity.Stock;
+import com.chuseok22.stocktrader.common.infrastructure.entity.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StockJpaRepository extends JpaRepository<Stock, Long> {
