+++ src/main/java/com/chuseok22/stocktrader/common/infrastructure/repository/UniverseItemJpaRepository.java
@@ -1,6 +1,6 @@
-package com.chuseok22.stocktrader.infrastructure.repository;
+package com.chuseok22.stocktrader.common.infrastructure.repository;
 
-import com.chuseok22.stocktrader.infrastructure.entity.UniverseItem;
+import com.chuseok22.stocktrader.common.infrastructure.entity.UniverseItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UniverseItemJpaRepository extends JpaRepository<UniverseItem, Long> {
