+++ src/main/java/com/chuseok22/stocktrader/common/infrastructure/repository/UniverseJpaRepository.java
@@ -1,0 +1,8 @@
+package com.chuseok22.stocktrader.common.infrastructure.repository;
+
+import com.chuseok22.stocktrader.common.infrastructure.entity.Universe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UniverseJpaRepository extends JpaRepository<Universe, Long> {
+
+}
