--- src/main/java/com/chuseok22/stocktrader/core/constants/Region.java
@@ -1,5 +1,0 @@
-package com.chuseok22.stocktrader.core.constants;
-
-public enum Region {
-  KR, US
-}
